+++ CHANGELOG.md
@@ -3,4 +3,4 @@
 ## [0.35.0]
 
 - Released @ 2024-05-24 02:32:01.155435Z
-- Changed DataRef to a Model
+- Changed DataRefModel to a Model